# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CineAI API
  description: Comprehensive movie recommendation API with AI-powered features
  version: 1.0.0
  contact:
    name: CineAI Team
    email: contact@cineai.app
  license:
    name: MIT
servers:
  - url: https://cineai.vercel.app/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Movie:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        overview:
          type: string
        posterPath:
          type: string
        backdropPath:
          type: string
        releaseDate:
          type: string
          format: date
        runtime:
          type: integer
        rating:
          type: number
          format: float
        voteCount:
          type: integer
        genres:
          type: array
          items:
            type: string
        cast:
          type: array
          items:
            type: object
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
...
paths:
...
  /watchlist:
    get:
      summary: Get user watchlist
      tags:
        - User Data
      security:
        - bearerAuth: []
      parameters:
        - name: watched
          in: query
          schema:
            type: boolean
          description: Filter by watched status
      responses:
        '200':
          description: Watchlist items
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                              format: uuid
                            movie:
                              $ref: '#/components/schemas/Movie'
                            watched:
                              type: boolean
                            rating:
                              type: integer
                            notes:
                              type: string
                            addedAt:
                              type: string
                              format: date-time

    post:
      summary: Add movie to watchlist
      tags:
        - User Data
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                movieId:
                  type: string
                  format: uuid
                source:
                  type: string
                  enum: [recommendation, search, manual]
              required:
                - movieId
      responses:
        '201':
          description: Movie added to watchlist
        '400':
          $ref: '#/components/responses/ValidationError'
...
security:
  - bearerAuth: []